Fixed a bug where the API returned truncated tweets, resulting in invalid wordle processing. Test id = 1608040651546836995 , reported here: https://twitter.com/thomasomatic/status/1608577819294183425
Fix was to include tweet_mode = 'extended' to all API calls, and to check the truncated flag on the tweet to determine whether to use the full_text or normal text property for processing. There are some other issues with invalid wordle checking in the processTweet function which are fixed in the refactored typescript version, coming soon.